# Conflicts:
#	backend/src/main/java/htlkaindorf/backend/repositories/UserRepository.java

+++ backend/src/main/java/htlkaindorf/backend/repositories/UserRepository.java
@@ -1,10 +1,14 @@
 package htlkaindorf.backend.repositories;
 
 import htlkaindorf.backend.pojos.User;
+import htlkaindorf.backend.pojos.UserSkin;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+import java.util.List;
+
+public interface UserRepository extends JpaRepository<User, Integer> {
     @Query("SELECT u FROM User u WHERE u.username = ?1")
     User findByUsername(String username);
+
 }
